follow unsigned change in dev-20200626

+++ bm_poly.cc
@@ -162,8 +162,8 @@ void EVAL_BM_POLY::tr_eval(ELEMENT* d)const
 /*--------------------------------------------------------------------------*/
 bool EVAL_BM_POLY::parse_numlist(CS& cmd)
 {
-  unsigned start = cmd.cursor();
-  unsigned here = cmd.cursor();
+  size_t start = cmd.cursor();
+  size_t here = cmd.cursor();
   for (;;) {
     unsigned old_here = here;
     PARAMETER<double> val;
